move index creation to DocumentStoreInitializer
