+++ src/NServiceBus.RavenDB/Internal/DocumentStoreInitializer.cs
@@ -1,10 +1,13 @@
 ﻿namespace NServiceBus.Persistence.RavenDB
 {
     using System;
+    using System.Collections.Generic;
     using NServiceBus.ConsistencyGuarantees;
     using NServiceBus.ObjectBuilder;
     using NServiceBus.Settings;
     using Raven.Client.Documents;
+    using Raven.Client.Documents.Indexes;
+    using Raven.Client.Documents.Operations.Indexes;
     using Raven.Client.ServerWide.Commands;
 
     class DocumentStoreInitializer
@@ -16,10 +19,39 @@
 
         internal DocumentStoreInitializer(IDocumentStore store)
         {
-            storeCreator = (s,c) => store;
+            storeCreator = (s, c) => store;
         }
 
         public string Identifier => docStore?.Identifier;
+
+        /// <summary>
+        /// Safely add the index to the RavenDB database, protect against possible failures caused by documented
+        /// and undocumented possibilities of failure.
+        /// Will throw iff index registration failed and index doesn't exist or it exists but with a non-current definition.
+        /// </summary>
+        internal void SafelyCreateIndex(AbstractIndexCreationTask index)
+        {
+            indexesToCreate.Add(index);
+        }
+
+        void CreateIndexes(IDocumentStore store)
+        {
+            foreach (var index in indexesToCreate)
+            {
+                try
+                {
+                    index.Execute(store);
+                }
+                catch (Exception) // Apparently ArgumentException can be thrown as well as a WebException; not taking any chances
+                {
+                    var getIndexOp = new GetIndexOperation(index.IndexName);
+
+                    var existingIndex = store.Maintenance.Send(getIndexOp);
+                    if (existingIndex == null || !index.CreateIndexDefinition().Equals(existingIndex))
+                        throw;
+                }
+            }
+        }
 
         internal IDocumentStore Init(ReadOnlySettings settings, IBuilder builder)
         {
@@ -30,7 +62,10 @@
 
                 docStore.Initialize();
                 EnsureClusterConfiguration(docStore);
+
+                CreateIndexes(docStore);
             }
+
             isInitialized = true;
             return docStore;
         }
@@ -81,6 +116,7 @@
             }
         }
 
+        List<AbstractIndexCreationTask> indexesToCreate = new List<AbstractIndexCreationTask>();
         Func<ReadOnlySettings, IBuilder, IDocumentStore> storeCreator;
         IDocumentStore docStore;
         bool isInitialized;
