+++ src/NServiceBus.RavenDB/Internal/Helpers.cs
@@ -3,8 +3,6 @@
     using System;
     using NServiceBus.Logging;
     using Raven.Client.Documents;
-    using Raven.Client.Documents.Indexes;
-    using Raven.Client.Documents.Operations.Indexes;
 
     class Helpers
     {
@@ -32,29 +30,6 @@ Original exception: {exception}";
             {
                 LogRavenConnectionFailure(e, documentStore);
                 throw;
-            }
-        }
-
-        /// <summary>
-        /// Safely add the index to the RavenDB database, protect against possible failures caused by documented
-        /// and undocumented possibilities of failure.
-        /// Will throw iff index registration failed and index doesn't exist or it exists but with a non-current definition.
-        /// </summary>
-        /// <param name="store"></param>
-        /// <param name="index"></param>
-        internal static void SafelyCreateIndex(IDocumentStore store, AbstractIndexCreationTask index)
-        {
-            try
-            {
-                index.Execute(store);
-            }
-            catch (Exception) // Apparently ArgumentException can be thrown as well as a WebException; not taking any chances
-            {
-                var getIndexOp = new GetIndexOperation(index.IndexName);
-
-                var existingIndex = store.Maintenance.Send(getIndexOp);
-                if (existingIndex == null || !index.CreateIndexDefinition().Equals(existingIndex))
-                    throw;
             }
         }
     }
